refactor(hydra): split snapshot verification by responsibility
snapshot-verification.ts was 840 lines, past this repo's 750-line
ceiling, and the 2.4.1 upgrade added to it. Extract the two parts it
grew: snapshot-serialization.ts for canonical Plutus TxOut bytes, and
snapshot-signable.ts for the bytes a party signs, which upstream dictates
verbatim. Both are re-exported, the way the accumulator split already
was, so no importer changes. 840 to 516 lines, no behaviour change.

Also correct the depositTxId comment. It said the field tightens the
transition check; nothing reads it. It is authenticated by being bound
into the signature and deliberately not acted on, which the transition
check itself explains at length.
